+++ codegpt-core/src/main/java/ee/carlrobert/embedding/EmbeddingsService.java
@@ -9,9 +9,9 @@ import com.github.jelmerk.knn.SearchResult;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.progress.ProgressIndicator;
 import ee.carlrobert.llm.client.openai.OpenAIClient;
-import ee.carlrobert.llm.client.openai.completion.chat.OpenAIChatCompletionModel;
-import ee.carlrobert.llm.client.openai.completion.chat.request.OpenAIChatCompletionMessage;
-import ee.carlrobert.llm.client.openai.completion.chat.request.OpenAIChatCompletionRequest;
+import ee.carlrobert.llm.client.openai.completion.OpenAIChatCompletionModel;
+import ee.carlrobert.llm.client.openai.completion.request.OpenAIChatCompletionMessage;
+import ee.carlrobert.llm.client.openai.completion.request.OpenAIChatCompletionRequest;
 import ee.carlrobert.splitter.SplitterFactory;
 import ee.carlrobert.vector.VectorStore;
 import ee.carlrobert.vector.Word;
@@ -64,16 +64,16 @@ public class EmbeddingsService {
   }
 
   public List<Item<Object, double[]>> createEmbeddings(
-      List<CheckedFile> checkedFiles,
+      List<ReferencedFile> referencedFiles,
       @Nullable ProgressIndicator indicator) {
     var words = new ArrayList<Item<Object, double[]>>();
-    for (int i = 0; i < checkedFiles.size(); i++) {
+    for (int i = 0; i < referencedFiles.size(); i++) {
       try {
-        var checkedFile = checkedFiles.get(i);
-        addEmbeddings(checkedFile, words);
+        var referencedFile = referencedFiles.get(i);
+        addEmbeddings(referencedFile, words);
 
         if (indicator != null) {
-          indicator.setFraction((double) i / checkedFiles.size());
+          indicator.setFraction((double) i / referencedFiles.size());
         }
       } catch (Throwable t) {
         // ignore
@@ -101,24 +101,26 @@ public class EmbeddingsService {
         .getContent();
   }
 
-  private void addEmbeddings(CheckedFile checkedFile, List<Item<Object, double[]>> prevEmbeddings) {
-    var fileExtension = checkedFile.getFileExtension();
+  private void addEmbeddings(
+      ReferencedFile referencedFile,
+      List<Item<Object, double[]>> prevEmbeddings) {
+    var fileExtension = referencedFile.getFileExtension();
     var codeSplitter = SplitterFactory.getCodeSplitter(fileExtension);
     if (codeSplitter != null) {
       var chunks = codeSplitter.split(
-          checkedFile.getFileName(),
-          checkedFile.getFileContent());
+          referencedFile.getFileName(),
+          referencedFile.getFileContent());
       var embeddings = openAIClient.getEmbeddings(chunks);
       for (int i = 0; i < chunks.size(); i++) {
         prevEmbeddings.add(
-            new Word(chunks.get(i), checkedFile.getFileName(), normalize(embeddings.get(i))));
+            new Word(chunks.get(i), referencedFile.getFileName(), normalize(embeddings.get(i))));
       }
     } else {
-      var chunks = splitText(checkedFile.getFileContent(), 400);
+      var chunks = splitText(referencedFile.getFileContent(), 400);
       var embeddings = getEmbeddings(chunks);
       for (int i = 0; i < chunks.size(); i++) {
         prevEmbeddings.add(
-            new Word(chunks.get(i), checkedFile.getFileName(), normalize(embeddings.get(i))));
+            new Word(chunks.get(i), referencedFile.getFileName(), normalize(embeddings.get(i))));
       }
     }
   }