+++ src/test/java/ee/carlrobert/codegpt/completions/CompletionRequestProviderTest.java
@@ -13,12 +13,10 @@ import ee.carlrobert.codegpt.conversations.ConversationService;
 import ee.carlrobert.codegpt.conversations.message.Message;
 import ee.carlrobert.codegpt.credentials.OpenAICredentialsManager;
 import ee.carlrobert.codegpt.settings.configuration.ConfigurationState;
-import ee.carlrobert.codegpt.settings.service.ServiceType;
 import ee.carlrobert.codegpt.settings.state.OpenAISettingsState;
-import ee.carlrobert.codegpt.settings.state.SettingsState;
 import ee.carlrobert.llm.client.http.ResponseEntity;
 import ee.carlrobert.llm.client.http.exchange.BasicHttpExchange;
-import ee.carlrobert.llm.client.openai.completion.chat.OpenAIChatCompletionModel;
+import ee.carlrobert.llm.client.openai.completion.OpenAIChatCompletionModel;
 import java.util.List;
 import java.util.Map;
 import testsupport.IntegrationTest;
@@ -36,9 +34,13 @@ public class CompletionRequestProviderTest extends IntegrationTest {
     conversation.addMessage(secondMessage);
 
     var request = new CompletionRequestProvider(conversation)
-        .buildOpenAIChatCompletionRequest(OpenAIChatCompletionModel.GPT_3_5.getCode(),
-            new Message("TEST_CHAT_COMPLETION_PROMPT"),
-            false,
+        .buildOpenAIChatCompletionRequest(
+            OpenAIChatCompletionModel.GPT_3_5.getCode(),
+            new CallParameters(
+                conversation,
+                ConversationType.DEFAULT,
+                new Message("TEST_CHAT_COMPLETION_PROMPT"),
+                false),
             false,
             null);
 
@@ -61,8 +63,15 @@ public class CompletionRequestProviderTest extends IntegrationTest {
     conversation.addMessage(secondMessage);
 
     var request = new CompletionRequestProvider(conversation)
-        .buildOpenAIChatCompletionRequest(OpenAIChatCompletionModel.GPT_3_5.getCode(),
-            new Message("TEST_CHAT_COMPLETION_PROMPT"), false);
+        .buildOpenAIChatCompletionRequest(
+            OpenAIChatCompletionModel.GPT_3_5.getCode(),
+            new CallParameters(
+                conversation,
+                ConversationType.DEFAULT,
+                new Message("TEST_CHAT_COMPLETION_PROMPT"),
+                false),
+            false,
+            null);
 
     assertThat(request.getMessages())
         .extracting("role", "content")
@@ -84,8 +93,15 @@ public class CompletionRequestProviderTest extends IntegrationTest {
     conversation.addMessage(secondMessage);
 
     var request = new CompletionRequestProvider(conversation)
-        .buildOpenAIChatCompletionRequest(OpenAIChatCompletionModel.GPT_3_5.getCode(),
-            secondMessage, true);
+        .buildOpenAIChatCompletionRequest(
+            OpenAIChatCompletionModel.GPT_3_5.getCode(),
+            new CallParameters(
+                conversation,
+                ConversationType.DEFAULT,
+                secondMessage,
+                true),
+            false,
+            null);
 
     assertThat(request.getMessages())
         .extracting("role", "content")
@@ -107,8 +123,15 @@ public class CompletionRequestProviderTest extends IntegrationTest {
     conversation.discardTokenLimits();
 
     var request = new CompletionRequestProvider(conversation)
-        .buildOpenAIChatCompletionRequest(OpenAIChatCompletionModel.GPT_3_5.getCode(),
-            new Message("TEST_CHAT_COMPLETION_PROMPT"), false);
+        .buildOpenAIChatCompletionRequest(
+            OpenAIChatCompletionModel.GPT_3_5.getCode(),
+            new CallParameters(
+                conversation,
+                ConversationType.DEFAULT,
+                new Message("TEST_CHAT_COMPLETION_PROMPT"),
+                false),
+            false,
+            null);
 
     assertThat(request.getMessages())
         .extracting("role", "content")
@@ -127,8 +150,15 @@ public class CompletionRequestProviderTest extends IntegrationTest {
 
     assertThrows(TotalUsageExceededException.class,
         () -> new CompletionRequestProvider(conversation)
-            .buildOpenAIChatCompletionRequest(OpenAIChatCompletionModel.GPT_3_5.getCode(),
-                createDummyMessage(100), false));
+            .buildOpenAIChatCompletionRequest(
+                OpenAIChatCompletionModel.GPT_3_5.getCode(),
+                new CallParameters(
+                    conversation,
+                    ConversationType.DEFAULT,
+                    createDummyMessage(100),
+                    false),
+                false,
+                null));
   }
 
   public void testContextualSearch() {
@@ -175,8 +205,15 @@ public class CompletionRequestProviderTest extends IntegrationTest {
     });
 
     var request = new CompletionRequestProvider(conversation)
-        .buildOpenAIChatCompletionRequest(OpenAIChatCompletionModel.GPT_3_5.getCode(),
-            new Message("TEST_CHAT_COMPLETION_PROMPT"), false, true, null);
+        .buildOpenAIChatCompletionRequest(
+            OpenAIChatCompletionModel.GPT_3_5.getCode(),
+            new CallParameters(
+                conversation,
+                ConversationType.DEFAULT,
+                new Message("TEST_CHAT_COMPLETION_PROMPT"),
+                false),
+            true,
+            null);
 
     assertThat(request.getModel()).isEqualTo("gpt-3.5-turbo");
     assertThat(request.getMessages().size()).isEqualTo(1);